[x64] Factor out a lot of the opcode handling code
